prov/sockets: Fix 32-build warnings and crash
The sockets provider generates multiple build warnings and crashes
when running on a 32-bit system.  Fix the warnings and adjust the
code to handle 32-bit pointers, rather than assuming that the size
of all pointers is 64-bits.

To fix the 32-bit issues required performing a fairly low-level
review of the code.  Optimistic updates were made to the code to
align with coding guidelines, improve readability, etc.

+++ prov/sockets/src/sock_dom.c
@@ -244,7 +244,7 @@ struct sock_mr *sock_mr_verify_key(struct sock_domain *domain, uint16_t key,
 struct sock_mr *sock_mr_verify_desc(struct sock_domain *domain, void *desc, 
 			void *buf, size_t len, uint64_t access)
 {
-	uint64_t key = (uint64_t)desc;
+	uint64_t key = (uintptr_t) desc;
 	return sock_mr_verify_key(domain, key, buf, len, access);
 }
 
@@ -291,7 +291,7 @@ static int sock_regattr(struct fid *fid, const struct fi_mr_attr *attr,
 	if (idm_set(&dom->mr_idm, key, _mr) < 0)
 		goto err;
 	_mr->mr_fid.key = key;
-	_mr->mr_fid.mem_desc = (void *)key;
+	_mr->mr_fid.mem_desc = (void *) (uintptr_t) key;
 	fastlock_release(&dom->lock);
 
 	_mr->iov_count = attr->iov_count;
